Horizontal and vertical SIMD matmul

+++ sling/myelin/graph.h
@@ -51,7 +51,7 @@ struct GraphOptions {
   bool include_intermediates = false;
   bool cluster_functions = true;
   int max_value_size = 8;
-  float edge_thickness_scalar = 0.0;
+  int edge_thickness_scalar = 0;
 
   // Options for operations, inputs, outputs, and constants.
   GraphNodeOptions ops;
